ExperimentingWithAutoGeneratedResponse
As I need a universal response that can handle both crypto and company stock, it would take me too much time to waste to build it myself so I used an auto generator for the request and I'll just have to do the mapping.
This is a mocked up version of the mapping.